+++ lib/src/resizable_widget_child_data.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ResizableWidgetChildData {
+  final Widget widget;
+  double? size;
+  double? percentage;
+  double? defaultPercentage;
+  double? hidingPercentage;
+  ResizableWidgetChildData(this.widget, this.percentage);
+}