+++ CHANGELOG.md
@@ -1,10 +1,18 @@
+## 1.0.5
+
+* Add "Smart-Hide-Function" that users can hide / show the widget on the end by double-clicking the separator.
+* Add `isDisabledSmartHide` that means to prohibit the above feature.
+* Deprecate `isColumnChildren` and create `isHorizontalSeparator` that has the same meaning.
+
 ## 1.0.4
+
 * Publish the typedef `OnResizedFunc` for the API document.
 * Change the argument name of `OnResizedFunc`, but the type is not changed.
 * Update some documentation comments.
 * Update README.
 
 ## 1.0.3
+
 * Add `onResized` callback to `ResizableWidget`.
 * Update README.